payment shipping done

+++ database/migrations/2021_04_17_062212_create_cities_table.php
@@ -15,7 +15,7 @@ class CreateCitiesTable extends Migration
     {
         Schema::create('cities', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('province_id');
+            $table->unsignedBigInteger('province_id');
             $table->unsignedInteger('city_id');
             $table->string('name');
             $table->timestamps();
